+++ scripts/ci_scripts.py
@@ -20,6 +20,7 @@ scripts = [
     "datatools_models.py",
     "datatools_modeler.py",
     "unit_tests.py",
+    "docx_writer.py",
 
     # Workflows
     "workflow/blocks.py",